[FLINK-4469] [table] Minor improvements
- Fixed typos
- Removed implicit conversion with TableCallBuilder
- Fixed bugs about expression parser alias and static eval methods
- Refactored tests

This closes #2653.
