+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/functions/utils/ScalarSqlFunction.scala
@@ -125,6 +125,7 @@ object ScalarSqlFunction {
     : SqlOperandTypeChecker = {
 
     val signatures = getSignatures(scalarFunction)
+
     /**
       * Operand type checker based on [[ScalarFunction]] given information.
       */
@@ -178,5 +179,4 @@ object ScalarSqlFunction {
       }
     }
   }
-
 }
