+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/functions/utils/TableSqlFunction.scala
@@ -33,7 +33,6 @@ import org.apache.flink.api.table.plan.schema.FlinkTableFunctionImpl
 import scala.collection.JavaConverters._
 import java.util
 
-
 /**
   * Calcite wrapper for user-defined table functions.
   */
@@ -55,31 +54,33 @@ class TableSqlFunction(
     functionImpl) {
 
   /**
-    * Get the user-defined table function
+    * Get the user-defined table function.
     */
   def getTableFunction = udtf
 
   /**
-    * Get the returned table type information of the table function
+    * Get the type information of the table returned by the table function.
     */
   def getRowTypeInfo = rowTypeInfo
 
   /**
     * Get additional mapping information if the returned table type is a POJO
-    * (POJO types have no deterministic field order)
+    * (POJO types have no deterministic field order).
     */
   def getPojoFieldMapping = functionImpl.fieldIndexes
 
 }
 
 object TableSqlFunction {
+
   /**
-    * Util function to create a [[TableSqlFunction]]
+    * Util function to create a [[TableSqlFunction]].
+    *
     * @param name function name (used by SQL parser)
-    * @param udtf user defined table function to be called
+    * @param udtf user-defined table function to be called
     * @param rowTypeInfo the row type information generated by the table function
     * @param typeFactory type factory for converting Flink's between Calcite's types
-    * @param functionImpl calcite table function schema
+    * @param functionImpl Calcite table function schema
     * @return [[TableSqlFunction]]
     */
   def apply(
